Possible fix for Lockup Caused in Noise Sensor at 2nd getReading
I took a clue from various reports of lockups of  I2S arduino libraries caused when I2S stop() function executed.

So, I moved begin() command to the ::start() function.
and moved the I2S.end() function call into the ::stop() function.

At first blush this appears to have fixed the lockup issue.

I am not clear about the ramifications of making this change; probably (at best) it is just a work-around.  But it works.

I also cleaned up the bit of code at the top of auxboards:getReading : This now needs to be retested.