Updated with latest changes
Updated today with changes that include:
(a) Handling of SCD41 at startup: removed pressure compensation:  The result is that CO2 readings are being returned.  However the values are not believable iro 17000 ppm.  This needs further work.
(b) changes made to UI control command for SCD41 allowing pressure compensation to be enabled/disabled;
(c) Changes for SckUrban to stop a hang when Noise sensor is used.  It was caused by a hang of I2S (begin) on second reading cycle.
The solution was to put the begin() command into ::start(). and the I2S.end(0 command into ::stop.  
(d) the beginning of changes to support 2nd OLED display.  This is WIP; its not working just at the moment.

There are a number of other bugs appearing in this version; detailed in the issues()